Updated per PR feedback.
