+++ .babelrc
@@ -2,6 +2,6 @@
   "presets": [
     "env",
     "react",
-    "stage-2"
+    "stage-0"
   ]
 }
